+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Parameter `-f` or `--force` in `vtex publish` to skip check for sem ver.
 
 ## [2.69.1] - 2019-09-04
 ### Fixed